+++ pkg/networkserver/networkserver_flow_test.go
@@ -923,7 +923,7 @@ func TestFlow(t *testing.T) {
 						}()
 					}
 
-					conf := DefaultConfig
+					conf := HandleDefaultConfigValues(DefaultConfig)
 					conf.NetID = test.Must(types.NewNetID(2, []byte{1, 2, 3})).(types.NetID)
 					conf.ApplicationUplinks = uq
 					conf.Devices = dr